[errors] Filter duplicate errors due to UnifyFlip during post inference
Summary:
During subtyping when we have to unify due to invariant subtyping rules, we will wrap the use_op with `UnifyFlip` frame. e.g. for `Array<string> ~> Array<number>`, we will emit the following two flows:

```
string ~> UseT (ArrayCompat(inner_use_op), number)
number ~> UseT (UnifyFlip(ArrayCompat(inner_use_op)), string)
```

If we don't do anything special, then we will emit two errors, which doesn't help users. Therefore, we have special logic to deduplicate errors with the `dedupe_flip` function. Intuitively, it will flip the error from the second flow into the same message as the first one, so when they are added to the set later, it will only be one error.

It's important to note that it's not not always the case that during unification failure, two errors will be emitted. For example, `Array<Dog> ~> Array<Animal>` will only emit error in one direction. This is the case I'm interested in, since I want to add some special explanations to it, so that the user can be aware it's due to invariant subtyping.

However, it's not enough to just match on `UnifyFlip(ArrayCompat)` and do something special. Doing so will cause us to emit two distinct error messages, and destroy the deduplication done by `dedupe_flip`. In this diff, I solve the problem by pre-removing `UnifyFlip` errors if there is already an existing equivalent error during the post inference pass, so `dedupe_flip` can be removed from `flow_error.ml`. This has an added benefit that the direction of type incompatibility error messages is finally correct. Instead of `Dog is incompatible with Aninal` in the reverse direction, we have the correct `Animal is incompatible with Dog`. The previous wrong direction is only there for us to dedupe the error messages.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision:
D43282936

------------------------------------------------------------------------
(from 28556b719655a278dcf9e117de7ee9abe970678f)

fbshipit-source-id: 59964d11c2dd9219f5fd9a16003d19e5d9c14b81