[conditional-type] Conditional type normalization
Summary:
Most of the changes are quite mechanical, except for the normalization part that decides whether a generic type should be normalized to infer type.

To achieve that, we keep a separate infer tparam set in the normalizer env. The set got replaced with a fresh set from conditional type destructor. When we are converting a generic type, we will check whether the generic id and name of the type matches anything in the infer type set.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D44151782

fbshipit-source-id: 710bdb0cba916d1264d7c44d58d70fe3b45f8290