Kill module name candidates cache
Summary:
The candidates cache is never cleared. It's technically bounded by the number of unique requires in a given root, but it's still wasted memory.

Do we really need this cache? It turns out: no. We only use this cache during resolve requires, and profiling shows absolutely no loss in perf, even for a full check which resolves every require in a given root.

We could optimize the module name mapping logic quite a lot, but it doesn't seem necessary or worthwhile, so I'm leaving this as-is.

Reviewed By: mroch

Differential Revision: D44451481

fbshipit-source-id: 0fbf22c1cbabd56078dd2dc0616b771639e83623