[parser] improve recovery of mismatched jsx tags
Summary:
if a JSX element's opening and closing tags don't match, and the closing tag matches the parent's opening tag, then we're in a situation like `<a><b></b><c></a>`, where we're parsing `c`:

`opening_name = c`
`closing_name = a`
`parent_opening_name = a`

we use this heuristic to assume that `c` is unclosed and we can recover by pairing up `<a>` and `</a>`.

i couldn't find a clean way to consume `</a>` while parsing `<c>` and return it to the parent, because of the varied places `child` and `element` are called recursively. so I just let it get parsed as the closing tag, and then the parent steals the closing tag from the child, so that `<c>` has no closing tag, but `<a>...</a>` is properly paired.

Changelog: [parser] Improve error recovery for mismatched JSX tags

Reviewed By: darichey

Differential Revision: D44462438

fbshipit-source-id: 13cac3a765eebf4e7b0b4ac400a5f0969352d2de