Control worker mode with flowconfig option
Summary:
Adds a new worker mode where jobs are handled inline instead of in a forked
process. Building on top of D44484163, the changes here are quite simple, but
the implications are pretty big:

On Windows, Flow needs to spawn a process for each job, which can be horribly
slow. With long-lived workers, we can handle jobs much faster.

Furthermore, this change moves us closer to worker domains in Multicore OCaml,
which would also not be able to simply exit like a forked process.

Compared to forking, we need to be careful about memory use. If the job has a
leak, the worker can eventually exhaust memory or be OOM killed. I have done
some memory profiling and fixed a number of existing leaks, but will also need
some better tooling to detect and prevent future leaks.

Even if the worker does not leak memory, OCaml heap blocks allocated during a
job will stick around between jobs. Even if the GC is keeping pace with
allocations, it's likely that idle memory use will be higher with long-lived
workers.

That said, this diff represents a working proof-of-concept, gated behind a flag.
We can experiment with long-lived workers locally.

Reviewed By: mroch

Differential Revision: D44484715

fbshipit-source-id: fd30916781a90489c0b44425122692b9d59fb54a