Unwind to worker main on worker cancel
Summary:
The server can tell workers to cancel by setting a flag in shared memory.
Workers periodically check this flag and raise an exception if it's on.

Before this change, we caught this exception and exited from the worker process
deep in the call stack, using `WorkerCancel.with_worker_exit`. This avoids a
potential bug where an try-with expression swallows the cancel exception.

We want to move to long lived workers, where the worker must not exit on cancel.
In order to provide this behavior, we will centralize the exception handling
code in the worker_main function instead.

As a first step, we will let the cancel exception unwind the stack all the way
back up to worker_main. Currently we still exit, but in later diffs I will add a
flag to simply continue the listen loop instead.

I audited the code for any catch-all try-with expressions. Happily, the only
ones I could find already handled similar unwind-to-main exceptions like
Out_of_shared_memory, so I only needed to add another case to existing code.

Reviewed By: mroch

Differential Revision: D44469303

fbshipit-source-id: 74a4684c766dbdf8a75bcf191927fabd3787dd1f