Consolidate worker_main exception handling
Summary:
We want to catch all exceptions at the top level of worker_main for logging. Any uncaught exceptions should make the worker exit, which will cause the server and other workers to also exit.

We also exit when a job completes or when a job is canceled. This diff combines those cases to a single `exit 0` at the end of the worker_main function. This makes it very simple to change the exit behavior. For long-lived workers, we want to continue instead of exit.

Reviewed By: mroch

Differential Revision: D44472628

fbshipit-source-id: 28284faf50bd74fa3f9cb31232ddc6667d1f441c