Refactor worker entrypoint to make room for long-lived workers mode
Summary:
Currently, the worker decides which entrypoint to use based on Sys.win32. It
either calls win32_worker_main which runs and exits, or unix_worker_main which
runs a listen loop and forks on each message.

This diff uses a single worker_main entrypoint which takes a worker_mode flag.
The flag is determined in ServerEnvBuild, where we have the `options` record
handy.

This diff also refactors the listen loop. Instead of hardcoding the fork
handler, the listen loop is parameterized by a handler. Under long lived
workers, the handler will simply be `worker_job_main`.

Reviewed By: mroch

Differential Revision: D44484163

fbshipit-source-id: c097b81bd301612aa7fceeae071a68a43348c892