Fix memory leak in const fold guard
Summary:
This module has a top-level cache which is never cleared. This diff moves the
cache to the context so it can be collected.

I also added this to the cache save/restore logic used for other caches.

This fixes some memory leaks that I observed while testing long-lived caches.

Reviewed By: mroch

Differential Revision: D44530811

fbshipit-source-id: 2981e07448c8ff2eaa02b9fe4abdd6a96ccaa909