Remove ProcFS-based memory profiling
Summary:
We use ProcFS profiling for two purposes:
1. To impose a memory limit for check jobs
2. For logging

Since the types-first 2.0 improvements to memory usage, the memory limit is not
really important. For logging, the information from ProcFS is misleading and we
use the information from cgroups instead.

Furthermore, with long-lived workers, this implementation of a memory cap
doesn't really make sense. With forked workers, ending a job causes the forked
process to die, and the next job would start out with lower memory use. With
long-lived workers, we rely on the GC to keep memory use in control.

Reviewed By: mroch

Differential Revision: D43142529

fbshipit-source-id: fbd9ec9e565e039268067272f70ea0278dd3e352