Fix memory leak in timer
Summary:
We were leaking the entry in the `cancelled` set when cancelling a timer that
was currently running.

The cancelled set is used in get_next_timer, to skip over cancelled timers when
they are dequeued.

However, the current_timer is by definition not in the pqueue, so we will never
reap the entry in the cancelled set.

The fix is simple: when cancelling the current timer, we do not need to add to
the cancelled set. Updating current_timer to None and re-scheduling is enough.

Reviewed By: mroch

Differential Revision: D44532142

fbshipit-source-id: d02842b9718bdebfbfea34333bf5b744bbabffb8