fix: catch SpeculativeError exceptions in implicit instantiation return hint flow
Summary:
In implicit instantiation, if there is a return hint available, we will first try to use it in order to help
type params be appropriately constrained. If this doesn't work, we roll back the state that was
accumulated during this flow, and redo the entire implicit instantiation assuming no return hint.

What we hadn't accounted for was the possibility that the return hint flow could raise an speculative
error exception. This was not a risk for safety: the exception would be propagated and cancel
the outer speculative branch. It did steal, however, the opportunity to try the implicit instantiation
without the return hint.

This diff adds the ability to recover from a SpeculativeError exception by a try-catch around the
return-hint flow. We treat this exception similar to any other potential errors that could arise during
that time (like we did before).

If we end up keeping the version of implicit instantiation that uses the return hint, we re-raise the
SpeculativeError exception.

Changelog: [fix] Fixed spurious error in overloaded generic calls when a contextual type is available (e.g. [try-Flow](https://flow.org/try#1N4Igxg9gdgZglgcxALlAJwKYEMwBcD6aArlLnALYYrgA2WAzvXGCADQgYAeOBARgJ74AJhhhYiNXClzEM7DFCLl602QF92kEdRFg6mAAQwSeONCMQIACgCUyAwAU0EcnHoYAPPRlwoCAHwA3AA6ULr6GAYAblhoBrz2vJY02FAhUKGQUN4GnAYAvPEGAPwGAMwG9lYwlrYAdGBYuGAAFlaR+f4GwGo2NoFsIFEYaEzQ1FEADHUATJNldZMgakA))

Reviewed By: SamChou19815

Differential Revision: D44550222

fbshipit-source-id: dcb4189ca67acbfc73b1c17c8a8ac64f06baa828