add useful message to poorly located errors
Summary:
All Flow errors have a so called primary location, which is the location that appears in the
error message, and is suppressible through `$FlowFixMe` comments. This location is typically
around the AST node that caused the Flow checker to raise an error. However, the logic
of determining this primary location is quite convoluted and is often seeded with unreliable
information from the checker. It is possible in certain cases for the primary location to be placed
far from the node that generated the error in the first place. In the extreme case, the error can
be located in a different file. This makes these errors extremely difficult to debug and sometimes
impossible to suppress (if the error appears to be in a core library definition).

This diff aims at easing the process of debugging such errors. For this kind of problematic errors
it will add the following text to the messages body:
```
(FLOW BUG: This is a misplaced error. The original error was raised in file `<remote file>`).
```
While these errors might still be impossible to suppress, this information makes it easier to fix
them instead.

This diff only targets Normal errors. The rules to determine misplaced Speculative errors (which
is the other kind of errors) are more convoluted and are left for future work.

The change also doesn't apply for libdef errors since the source file is arbitrarily chosen there.

Changelog: [fix] Add useful information to poorly located error messages. For example, for code like this [try-Flow](https://flow.org/try/#1N4Igxg9gdgZglgcxALlAJwKYEMwBcD6aArlLnALYYrgA2WAzvXGCADQgYAeOBARgJ74AJhhhYiNXClzEM7DFCLl602QF92kEdS4AHCGlwACXP10YjARSJZDGNEYC8RgCQAlbEIDyUGvwA8wGoAfADcADpQImB0mEYAbrZGAI42dmj0yEYAgmhoWAHWtrj2YZGRNBjG9AYlQkXp9AAycPTGzl68AFYYeAB0iTREGPQAFKnF9vQAlH0wdLij0xFQNXb1aSUZLW1ORgDafUdrdQ1bza24ALp9J6ORRuO4AIysKbgATNNOwQ9G-8kXrdIOZ8FAsJQ+jQIGAsJUAMIQci6WwYJ4fYEQUHgyjfAA+ePezz6cCiXCMAFp3hjSSJOKxIss2CB4lM4NBqPEAAx9D5cgDMfS5IDUQA) where the error points to a library definition, the error message will show the file that caused the error.

Reviewed By: SamChou19815

Differential Revision:
D42786670

------------------------------------------------------------------------
(from 966e15f7f653ec6c05ee8308b3c9400a5a42b591)

fbshipit-source-id: f0ffc6df339995e69f709421da1078da5eb27a32