Move parse options into its own module and rename to parsing options
Summary:
Whenever we add new typing syntax we need to interpret it in two places:
1. type_annotation.ml
2. type_sig_parse.ml

Importantly, type_sig_parse does not have access to options set in the flowconfig. If we return any for new experimental features when
the flag is disabled in type_annotation and do not do the same in type_sig_parse then we can get inconsistent behavior for
disabled experimental features depending on if they are used in a downstream import or in the file they are used.

The goal of this stack is to make it easier to add new options into Options.t and have them convert cleanly into Type_sig_options.t (introduced in the next diff, replacing Type_sig_parse.options).

When we create Type_sig_parse options we have one of two different options types on hand:
1. Parsing_service_js.parse_options
2. Options.t

This diff separates out Parsing_service_js.options into its own module so that we can define an `of_parsing_options` function in `type_sig_options.ml` that will be used in `parsing_service_js.ml` to convert parsing options into type sig options. Without separating out the module we have a cycle.

In addition, I'm renaming this parsing options to distinguish from Parser_env.parse_options.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D44344302

fbshipit-source-id: 0bd00848398af6ffacc520f2ffdc6d47f48aba87