Fix signatureHelp docs for member callees
Summary:
Previously, signature help didn't show documentation in situations like `o.m(|)`, because the callee loc was that of the entire `o.m` expression (and passing that to `Find_documentation` failed). Now, if the callee is a member expression, we use the location of the property for the callee loc.

Changelog: [ide] Fixed showing documentation in signatureHelp for member expressions (e.g., `o.m(|)`)

Reviewed By: mroch

Differential Revision: D44585125

fbshipit-source-id: 5a4375ee05ab71ff177a998c3ee728fd50ed623b