Fix autocomplete for super members
Summary:
In the typed ast, a `super` node is annotated with a type wrapped in a `ThisClassT`. By default, this causes the normalizer to look at its static members, but this is incorrect for members of `super`. If we are accessing a member on `super`, force the normalizer to examine its instance properties.

Changelog: [ide] Fixed autocomplete for super members (e.g., `super.|`)

Reviewed By: mroch

Differential Revision: D44589688

fbshipit-source-id: f91d034fb76fbb0d1a4a6147800aa82e445d05ee