[parser] minor fix in annotation parser
Summary:
In the expression
```
(Missing (Peek.loc_skip_lookahead env), Some (no_annot_predicate env))
```
the order of execution of the two subexpressions is not specified. In the unit tests with the error diff, the second
part was executed first, which doesn't seem like the wanted behavior.

This diff pulls out the subexpressions and assigns them first into local vars to avoid the unspecified behavior.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D44598792

fbshipit-source-id: 3af460e321e2408ac3656bc040fe04542eaa3a02