[conditional-type] Allow duplicate infer names
Summary:
We model infer type as tparams hoisted from extends type, so we must consider the case when there are name collisions. Unlike actual tparams, there are actual use cases. e.g. in `[string, string] extends [infer Y, infer Y] ? Y : empty`, it's used to express the constrain that both elements of the tuple type must have the same type.

This will affect how we implement these infer types. Previously, we generate a tparam for each infer type at the location of the infer type, and then hoist them to form a tparam list for the true branch. In the world where we allow duplicate infer names, this will no longer work, because then the GenericT for each infer type will refer to different infer types.

Therefore, in type_annotation.ml, we must first hoist all the infer types in extends type before checking it, do the deduplication, and then pass down the correct information to infer types so that it's referring to the correct tparam. In type_annotation.ml, this requires us to propagate infer_tparams_map everywhere.

With duplicate infer types, we need to also worry about conflicting bounds. I will let the first bound win in case of conflicts, and do `Flow.unify` to ensure that the bounds are consistent.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D44523083

fbshipit-source-id: 9919aa19a5b47ddb39d1f703a80a5a05b19e5418