[conditional-type] Fix generics substitution issue with an alpha rename pass during construction
Summary:
Right now, substitution with conditional type with infer tparams are quite broken. The infer tparam makes the part of the destructor more PolyT like, but we are not doing any alpha renaming for these nodes at all. Therefore, we need to first perform the avoid_capture setup like we do for PolyT.

Even with avoid_capture, it's still wrong, because we allow the infer type and GenericT with the same name, both implemented as GenericT, to coexist within extends_t of conditional type.
e.g.

```
type InferTypeOverlapGenerics<T> = {foo: T, bar: number} extends {foo: T, bar: infer T} ? T : boolean;
```

We have the info to distinguish between these two cases with generic id. However, our current substitution scheme is keyed only by Subst_name, so a naive implementation will let either ordinary tparam or infer tparam win, which is incorrect. e.g. when `T=string` is substituted in the example above, if we let the ordinary tparam win, then it will be `{foo: string, bar: number} extends {foo: string, bar: string} ? T : boolean`, which will cause us to choose the wrong branch!

So what should be the solution? Introducing a new type that just represent the infer type can fix the distinguishability issue, but then we have to worry about its interactions with all the other types. In additional, an infer type on its own is meaningless. Banning this in syntactic analysis might be appealing, but it's inconsistent with other parts of the type system when we allow tparam shadowing, so I did a more involved fix here.

I fixed the above mentioned problem by doing an alpha rename for infer tparams at the construction time. Then later when the internal type language of conditional type interact with the rest of the types, we have a strong guaranteed that an infer tparam will **never** share the same name with other upper tparams in scope. This allows a much easier substitution schema with conditional type everywhere else.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D44588940

fbshipit-source-id: 3595f5316a99c2c5243415be35dd4bdec1d773e8