Change multi-worker API to map-reduce
Summary:
Before this change, the multi-worker API seemed like a fold, where the job accepted an accumulator parameter, but it is not truly a fold. We don't want to pass an accumulator back to workers -- the communication overhead would be too much -- so instead we always just passed the "neutral" value.

Without any loss in generality, we can instead think of the multi-worker API as a map-reduce. The map function happens remotely, the reduce function happens locally.

This diff changes the "default" API, `call`, to the map-reduce form. We can see the benefit by looking at the callers also updated in this diff.

Many jobs are themselves a fold over a list. That is, the map function takes a list as input and folds over that list. For these callers, the map-reduce API is more verbose than the previous one. This diff keeps a "fold" like API, but specializes it to lists. This is an improvement over the existing API as well.

Finally, there is a single caller which has a "iter" like API. It is trivial to add this, so it's done here as well.

Reviewed By: mroch

Differential Revision: D44641581

fbshipit-source-id: 92426026432223297b53e12382ad2f1543c95b35