Make tvar graph more mutable
Summary:
For unresolved tvars, we mutate the underlying tvar constraints directly. For
other changes, we change the graph by creating a new node and overwriting the
IMap.t. This overwriting is needlessly expensive: each update is O(logn)
allocations because the spine of the graph needs to be updated.

This diff adds more interior mutability to the tvar graph, to support the
following changes:

1. Changing a Root to a Goto during unification
2. Changing Unresolved to FullyResolved
3. Flattening a chain of Goto nodes by updating the parent pointer
4. Incrementing the rank of a root node

Changelog: [internal]

Reviewed By: jbrown215, panagosg7

Differential Revision: D44675763

fbshipit-source-id: dca141c74bb1304456322e8c972c5276dea6aa28