Remove unnecessary tvar unification in annotation inference
Summary:
During annotation inference, we will create an unresolved avar and then resolve it to some type. In theory, there is a possible where the resolved-to type is itself unresolved. Before this diff, in this case, the code would attempt to unify the avars. This unification mirrors the behavior for tvars.

This diff takes a different approach: we instead insist that the resolved-to type is itself fully resolved. This eliminates the unification case entirely, which is a nice simplification which is explored more fully in stacked diffs.

Note that this change also implies that tvars are not unified, since the avars are 1:1 with tvars during annotation inference.

If the resolved-to type is not fully resolved, then it is a trivially recursive definition -- an error. To raise the error in the right place, this diff needs a reason, so the reason is threaded around.

Reviewed By: panagosg7

Differential Revision: D44655880

fbshipit-source-id: f111497617778c73231a8d4d83d1041bfea8e127