Treat missing avar as resolved types
Summary:
Before this diff, if we looked for an avar by ID and failed to find one, we would install a fully resolved avar. During annotation inference, this case only happens for types which come from dependencies, which are fully resolved by definition.

If we see a tvar id which has no corresponding avar id, then the tvar must be fully resolved. This diff exploits that fact by treating missing as fully resolved *without* adding an avar to the graph.

This change reduces the size of the avar graph and therefore improves performance.

Reviewed By: panagosg7

Differential Revision: D44657747

fbshipit-source-id: 283b5b56b9c57d9813a2431357123d5c1ced0d35