fix statics of template function in hint decomposition
Summary:
In hint decomposition we use function templates to match existing types in order to extract
various parts of a function type. The statics of this function template type have not been
important so far, but in the next they will become more important when this template is compared
against another object type.

To avoid errors that may throw off hint decomposition it is important to get the statics of this template
function right. This diff fixes the statics type for this function making it an object type with `FunProtoT`
as its prototype and with a "literal" reason, to relax checking.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D44740704

fbshipit-source-id: d684c48af050d91defc384c7f9d8c37cd0ac9ce1