Do not allow repositioning to stall implicit instantiation constraints
Summary:
Repositioning a tvar adds a ReposLowerT upper bound to the tvar and waits for the tvar to concretize. Once the tvar concretizes, the concrete lower bound is repositioned and flowed once more to the upper bound.

This is good for value-like tvars, which are the vast majority of the tvars that run through flow. However, implicitly instantiated tvars do not always receive lower bounds, so we should not expect them to eventually concretize. In many cases, it's possible that waiting for the tvar to concretize can block the tvar from receiving the bound it would get if repositioning was not implemented in this way.

This diff stops repositioning instantiable tvars during implicit instantiation. The implicit instantiation check is necessary because we sometimes create tvars with instantiable reasons outside of implicit instantiation, and the behavior for those cases is better with the repositioning than not. In particular, not repositioning those cases can lead to poorly positioned duplicate errors (as can be see in V1 of this diff).

Reviewed By: samwgoldman

Differential Revision: D44727765

fbshipit-source-id: d75329e846d3dad06f477379953c484bd1403be8