Fix get_def for constructors
Summary:
Try to get the "constructor" member of the callee for get def in `new` expressions.

Since "constructor" is always on the prototype chain for classes, by default pointing to the class declaration loc, we get fallback to that loc for free!

This also fixes showing constructor documentation on hover.

Changelog: [ide] Fix go-to-definition for constructors (e.g., `new Foo()`)

Reviewed By: mroch

Differential Revision: D44585123

fbshipit-source-id: f359deefe2f0abc4c9b59d9fffdab81ee6fd41bd