Use Ty_members for non-module get_defs
Summary:
Use `Ty_members` for `ObjectType` get def requests. Notably, this gives the option to use the `force_instance` functionality of `Ty_members` which I will need in the next diff.

Some extra changes mirroring those in D44589688 were needed to keep get def of `x` in `super.x` working properly.

I tried replacing `Members` entirely, but it didn't work for the `ObjectRequireLoc` code path, so I'm leaving that as-is for now.

Reviewed By: mroch

Differential Revision: D44585128

fbshipit-source-id: 38084e8c5d467789fd242be83225806c45a12f5c