check function static props when comparing with callable instance types
Summary:
Similar to previous diff but for constraints of the form `FunT _ ~> InstanceT { call_t = Some _ }`.
If the instance is a class instance, this constraint will always lead to an error. In the case of an interface
it will do property-wise checking between the function's statics and the instance type.

Changelog: [fix] When considering a function as a callable instance type, we now also check the statics of the function against the instance. This may cause new errors. For example, the code in this [try-Flow](https://flow.org/try/#1N4Igxg9gdgZglgcxALlAJwKYEMwBcD6aArlLnALYYrgA2WAzvXGCADQgYAeOBARgJ74AJhhhYiNXClzEM7DFCLl602QF92kEdRgk8caAAIYACgCUh4GoA6sAHQwIEQwF5DARgDct2yZjJDOFIMNDEwDEtzAIA3CDghT2MnAPoZIIQ1M082EGiQpmhqaIAGOwAmYoBmO3cQNSA) will now error.

Reviewed By: gkz

Differential Revision: D44758932

fbshipit-source-id: d62fc8e555602f875b697fe468fa3e10fed7d8f8