Change check_file to accept resolve_module instead of requires
Summary:
Currently, check_file callers construct the "requires" list, based on the file
sig, which contains resolved modules.

Later in this stack, when we parse the file sig inside of check_file, callers
will no longer have access to the file sig.

To prepare for this, we change the API here, so callers do not need access to
the file sig. In my opinion, this change independently makes the API a bit
nicer.

Reviewed By: panagosg7

Differential Revision: D44816018

fbshipit-source-id: 2afbfcbe4918e427c67bf61b0e0197b1693da213