Use requires, not file sig, to ensure checked dependencies for check-contents
Summary:
When we do check-contents, including for IDE commands, we first check that
dependencies have been checked. This is important for lazy mode, where a file's
dependencies might not have been parsed yet, so we don't have signature
information for them.

Before this diff, we used the file sig to get the list of requires. This diff
changes that to use "requires" instead. In future diffs, we will stop storing
this information during parse at all, so this change unblocks that later
change.

Reviewed By: mroch

Differential Revision: D44898554

fbshipit-source-id: 41e29de2d07a64ae749d5c4cc9757ebe8f421fd8