Look up declare module imports directly from builtins
Summary:
We support import statements in declared modules, so declared modules can depend
on each other. These imports can only resolve to other declared modules.

Before this diff, we resolved these imports using file sigs. The file sig stores
all requires in declared modules and we query builtins before checking, when
adding "require tvars."

This diff moves the logic out of file sig and into the env resolver and
statement passes. We keep track of whether we are in a declare module. If we
are, then any import resolves to a builtin module. Otherwise, we look at the
require map.

Note that we need to keep track of the declare module context in the name def
graph because env resolution happens in "demand" order according to uses, not as
an AST visitor.

Reviewed By: SamChou19815

Differential Revision: D43024778

fbshipit-source-id: 554b727941cdb7ac987b1b2b0e587b371e257b6e