Add component syntax to AST
Summary:
This diff adds the AST nodes for the component syntax. It only handles the keyword `component` right now and accepts standard function parameters. This will be updated to use the component parameter syntax in a future diff.

Right now there is no way to create or interact with these nodes as the parser has not been updated yet.

changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D44512586

fbshipit-source-id: 3a8933a39efdb162b2be2cf249c15a9b1e2745d0