[flow/glean] full symbol span in DeclarationLocation
Summary:
Motivation:  codemarkup/glass expects DeclarationLocation to contain the span of a whole declaration, not just the symbol.

This span isn't available directly from the ast, but following mroch 's suggestion, I used `DocumentSymbolProvider.provide_symbol_information`  to get this info.

To summarize, this diff
- updates the schema to add the full span in the Declaration* predicates
- adds the "selectionRange" field  (i.e the symbol span) to SymbolInformation, so we can build a map `symbol span -> full span`
- in gleanRunner, create a map `symbol span -> full span` and use it whenever we create a declaration to get the full span.

# Note

Other attempt in D44113259, which adds the span to the `Declarations` rather than the `DeclarationInfos`. But I think the approach in this diff is better. Declaration predicates act as keys and shouldn't carry the extra span information.

Downside, some Declaration don't have a corresponding DeclarationInfo, so I implemented a fallback in the codemarkup in the meantime.

Reviewed By: donsbot

Differential Revision: D44941856

fbshipit-source-id: 3cae3addd9b1b2ce7ccc2165f83185068b3e29bf