prepare AST for type-guards
Summary:
This diff makes two changes in the AST structure in preparation for adding type-guard nodes in the
following diff. These changes will only allow type guards `x is T` to appear as:
* the return type of a function type, or
* the return type of a function expression.

So in the next diff, type-guards will become variants of the `Ast.Type.Function.return_annotation`
and `Ast.Function.ReturnAnnot.t`, respectively.

An alternative to this decision could have been to allow type-guards it be part of `Ast.Type.t`. This
approach would require a fair bit of checking and error handling to happen in various parts of the
checking process. By restricting the AST appropriately, we're paying some of this cost upfront, but
enjoy a much more structured AST afterwards.

Most of the changes in this diff were mechanical. One complication was that since we're replacing
some instances of `Ast.Type.annotation_or_hint` with `Ast.Function.ReturnAnnot.t`,
we also need to override `function_return_annotation` in places where currently
`type_annotation_hint` is overridden.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D44590619

fbshipit-source-id: a16a133380dd7e64842988d6c83f23118d8ab164