type guard AST node
Summary:
Adds a type guard AST node as:
* the return type for functions types (`Ast.Type.return_annotation`)
* the return annotation of functions (`Ast.Function.ReturnAnnot`)

and propagates the changes throughout the codebase. These annotations are still not parsed so the
nodes are never constructed.

In most places a type guard is treated like an annotation (similar to `Annotated t` node). For typing
purposes it is equivalent to a boolean for now. This diff does not add any refining support.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D44593349

fbshipit-source-id: f6f85bcbf94878c70b199a5da4e125f30acb2d5b