[parser] parse colon early in function return annotation
Summary:
In order to parse type guards in the return annotation part of a function, we need to be able to peek ahead
3 tokens: `:`, `x` (param) and `is`. This is not possible in the parser, since we can only peek two tokens ahead,
for performance reasons. To work around this, after we've detected the colon, we can parse it, and then
branch based on the next two tokens (this part is coming up).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D44892899

fbshipit-source-id: cdbbcafdd7fe419ceb936b6167624cdb637a8a68