adds optional 'asserts' to TypeGuard node
Summary:
In the future we might want to support the forms
```
function f1(x: any): asserts x { throw "" };
function f2(x: any): asserts x is T { throw "" };
```
with semantics that assert that the type of the argument to each function is truthy or
number respectively. The intention is that the function will throw if `x` is *not* of that type.

This is the equivalent of `invariant(<predicate_expr>);`, but with a type-guard instead of
a predicate (expressed in the form of a conditional expression).

The AST changes are the addition of an "asserts" boolean field, and making the type part optional.
The combination of `asserts = false` and `t = None` is unreachable through parsing.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D45004905

fbshipit-source-id: cd7301e6ee2ea9e8ad61ece0f60699b4e54cf8d5