Deduplicate unify flip for EPropNotFound errors
Summary:
This diff extends the support for unify flip deduplication to EPropNotFound errors.

D43282936 first changed the mechanism of the deduplication from doing it on printed errors to doing it on Error_message. As a result, some errors that are previously correctly de-duplicated are now showing up as different errors.

e.g. During `Array<{foo: string}> ~> Array<{}>`, we will get two prop-missing errors due to invariant subtyping. Since we don't do the deduplication for EPropNotFound error, both errors will show up, and the one with UnifyFlip frame will have an additional `Arrays are invariantly typed` explanation.

This diff does the deduplication. It's simpler than others, because for this error we don't need to flip upper and lower.

Changelog: [errors] Some duplicate errors with `Arrays are invariantly typed` explanations are now correctly removed.

Reviewed By: panagosg7

Differential Revision: D44968218

fbshipit-source-id: 8e89662ac46d48e7afdc8421ed19e10b46bf2fcd