Fix recursion limit exceeded errors encountered while resolving very large unions
Summary:
ResolveUnionT resolves each member of a union, evaluating any EvalTs (and other types, like TypeAppTs) along the way.

When unions are very large, this can easily exceed the default 10k recursion limit. Bumping the recursion limit is not a good fix on its own because that just delays the next time this happens. Instead, we should use a new trace for each member we resolve so that resolution isn't as "deep" as the size of the union.

Reviewed By: SamChou19815

Differential Revision: D45152130

fbshipit-source-id: c92f64c9b0296858c6203039421c42628c9e61b7