log session length and token length in autocomplete responses
Summary:
an autocomplete session is a series of requests on the same token. we track how many requests, and how many characters, per session.

we currently only log this via the command that accepts a completion. when accepted, we know how many requests it took and how many characters were typed (the difference between them accounts for copy/paste and correcting typos).

this diff adds the same info to the logs for each response, not just the accepted ones. i'm primarily interested in acceptance rate: # acceptances / # of sessions. we can use this new logging to tell the # of sessions by counting the completion requests with session_requests = 1.

(note: in open source, these logs are unused)

Changelog: [internal]

Reviewed By: darichey

Differential Revision: D45217613

fbshipit-source-id: 4fd077d1132b6321eb946c3929670a95399da8dc