Remove catch-all exception handler from module resolution
Summary:
This was added in D42456384 to help debug a sharedmem heap corruption issue,
which we since found and fixed.

We rely on catching certain exceptions at the top-level of workers, notably
Worker_should_cancel. This handler causes the pattern match at the worker level
to fall through and cause crashes.

In addition to Worker_should_cancel, we also catch out-of-shared-memory
exceptions, which would be similarly affected, but are unexpected, whereas
worker cancelation is relatively very common.

Alternatively, I could have added cases for the exceptions we want to propagate,
but since we fixed the bug this code was hunting for, I lean toward deleting it
instead. Happy to change this, if requested.

Reviewed By: mroch

Differential Revision: D45243861

fbshipit-source-id: 89e637c96b828949df033797daaa535bd39a2c14