Better error positioning for bad required/partial
Summary:
We shouldn't propagate the use_op to the tout in Partial/Required, otherwise we can get poorly positioned errors. The upper bound has a better use_op associated with it that we should use instead. This matches the behavior in the read-only case.

Changelog: [bugfix] Fixed cases where uses of Partial/Required caused errors to be positioned poorly.

Reviewed By: gkz

Differential Revision: D45188921

fbshipit-source-id: 46cc9428b2fb397c4531b52d42cb3dbfa2594997