log ac_type with each autocomplete acceptance
Summary:
we classify the "type" of autocomplete we're doing based on the AST. for example, "Acid" (autocompleting an id) or "Acmem" (autocompleting a member expression) or "Ac_jsx_element" (autocompleting a JSX element name).

this is included in our internal telemetry for each `textDocument/completion` response, but wasn't included in the logging callback when a completion is accepted. this prevents me from determining an acceptance rate for different "types" of autocomplete. for example, I'm interested in whether member completions (which are very accurate) are accepted more than identifiers (which are way fuzzier).

so this diff threads around the `ac_type` so it can be included in the logging command of each completion item.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D45276881

fbshipit-source-id: e71c7e1b0801e650bb0e2a96da66c6cd6e0746b2